# Add json path to parse error messages