--- a/tests/Properties.hs
+++ b/tests/Properties.hs
@@ -49,28 +49,28 @@ toParseJSON :: (Arbitrary a, Eq a, Show a) =>
                (Value -> Parser a) -> (a -> Value) -> a -> Property
 toParseJSON parsejson tojson x =
     case parse parsejson . tojson $ x of
-      Error msg  -> failure "parse" msg x
-      Success x' -> x === x'
+      Error path msg -> failure "parse" (formatError path msg) x
+      Success x'     -> x === x'
 
 roundTrip :: (FromJSON a, ToJSON a, Show a) =>
              (a -> a -> Property) -> a -> a -> Property
 roundTrip eq _ i =
     case fmap fromJSON . L.parse value . encode . toJSON $ i of
-      L.Done _ (Success v) -> v `eq` i
-      L.Done _ (Error err) -> failure "fromJSON" err i
-      L.Fail _ _ err       -> failure "parse" err i
+      L.Done _ (Success v)      -> v `eq` i
+      L.Done _ (Error path err) -> failure "fromJSON" (formatError path err) i
+      L.Fail _ _ err            -> failure "parse" err i
 
 roundTripEq :: (Eq a, FromJSON a, ToJSON a, Show a) => a -> a -> Property
 roundTripEq x y = roundTrip (===) x y
 
 toFromJSON :: (Arbitrary a, Eq a, FromJSON a, ToJSON a, Show a) => a -> Property
 toFromJSON x = case fromJSON (toJSON x) of
-                Error err  -> failure "fromJSON" err x
-                Success x' -> x === x'
+                Error path err -> failure "fromJSON" (formatError path err) x
+                Success x'     -> x === x'
 
 modifyFailureProp :: String -> String -> Bool
 modifyFailureProp orig added =
-    result == Error (added ++ orig)
+    result == Error [] (added ++ orig)
   where
     parser = const $ modifyFailure (added ++) $ fail orig
     result :: Result ()