Every time you parse some JSON and get error message, it is very painful to track where the error is because existing error messages do not point to the place in JSON where parse error occured. JSON *syntax* errors are reported from Attoparsec parser and it gives position. But JSON *conversion* errors have no support for tracking position/context so quality of error messages is poor.

This change adds json path to error message:
1. Parser tracks current JSON path (in reverse order)
2. An `<?>` operation is added to annotate Parser with JSON context (JSON path). This operation is named after Attoparsec's `<?>` which serves similar purpose.
3. Built-in FromJSON instances are updated to provide context information
4. `.:` and `.:?` operations automatically add context information.

Context is reported in form of [JSON-Path](https://code.google.com/p/json-path/) expression e.g. `$.store.book[0].title`:
```
> eitherDecode storeJsonString :: Either String Store
Left "Error in \"$.store.book[0]\": key \"title\" not present"
```

Caveats:
1. There is no tests for error message yet. I believe those should be done in normal xUnit assertion style instead of QuickCheck property style.
2. Only first error is reported. Further work can include supporting collection of error messages and reporting all of them at once. This can be useful e.g. for validating HTTP JSON requests and reporting all errors at once for displaying them to user on a web form.