@zudov 2015-02-07
+1
Lack of descriptive error messages is often a pain.
Would like to see it merged.