@maximkulkin 2015-02-26
Please note: major2 version will need to be incremented because this patch introduces backward-incompatible changes