@maximkulkin 2015-05-25
Rebased to latest master