@bos 2015-05-25
Overall, I think I like this.

You can see that I've looked through the code and have asked for a number of insignificant formatting changes.

Before you go tweaking the formatting, please measure the impact on performance.

To do that, run some of the benchmarks before and after your change, and post a comment here (and update the commit description) to describe what difference you measure:

```
cd benchmarks
cabal sandbox init
cabal install -j --only-dependencies
cabal configure
cabal build -j
dist/build/aeson-benchmark-compare-with-json/aeson-benchmark-compare-with-json -m glob '**/*aeson*'
```