@maximkulkin 2015-05-25
Benchmark without patch:
```
benchmarking decode/en/aeson
time                 1.793 ms   (1.762 ms .. 1.824 ms)
                     0.997 R²   (0.994 R² .. 0.998 R²)
mean                 1.797 ms   (1.771 ms .. 1.832 ms)
std dev              98.14 μs   (74.03 μs .. 138.2 μs)
variance introduced by outliers: 39% (moderately inflated)

benchmarking decode/jp/aeson
time                 2.475 ms   (2.428 ms .. 2.511 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 2.550 ms   (2.516 ms .. 2.602 ms)
std dev              133.7 μs   (93.08 μs .. 202.7 μs)
variance introduced by outliers: 36% (moderately inflated)

benchmarking encode/en/aeson-to-bytestring
time                 413.0 μs   (403.6 μs .. 424.0 μs)
                     0.996 R²   (0.994 R² .. 0.999 R²)
mean                 408.3 μs   (403.3 μs .. 414.8 μs)
std dev              18.40 μs   (14.98 μs .. 24.01 μs)
variance introduced by outliers: 39% (moderately inflated)

benchmarking encode/en/aeson-via-text-to-bytestring
time                 1.073 ms   (1.054 ms .. 1.097 ms)
                     0.995 R²   (0.990 R² .. 0.998 R²)
mean                 1.067 ms   (1.052 ms .. 1.091 ms)
std dev              59.81 μs   (40.56 μs .. 95.40 μs)
variance introduced by outliers: 45% (moderately inflated)

benchmarking encode/en/aeson-to-text
time                 929.9 μs   (911.6 μs .. 948.0 μs)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 926.7 μs   (913.2 μs .. 942.7 μs)
std dev              51.20 μs   (42.51 μs .. 62.28 μs)
variance introduced by outliers: 46% (moderately inflated)

benchmarking encode/jp/aeson-to-bytestring
time                 410.2 μs   (405.5 μs .. 414.7 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 409.5 μs   (404.6 μs .. 414.4 μs)
std dev              16.14 μs   (13.74 μs .. 18.77 μs)
variance introduced by outliers: 34% (moderately inflated)

benchmarking encode/jp/aeson-via-text-to-bytestring
time                 1.095 ms   (1.069 ms .. 1.118 ms)
                     0.996 R²   (0.994 R² .. 0.997 R²)
mean                 1.099 ms   (1.080 ms .. 1.126 ms)
std dev              74.67 μs   (57.21 μs .. 112.3 μs)
variance introduced by outliers: 55% (severely inflated)

benchmarking encode/jp/aeson-to-text
time                 996.9 μs   (947.6 μs .. 1.039 ms)
                     0.989 R²   (0.984 R² .. 0.994 R²)
mean                 936.3 μs   (919.3 μs .. 958.8 μs)
std dev              64.58 μs   (51.88 μs .. 82.76 μs)
variance introduced by outliers: 56% (severely inflated)
```

Benchmark with patch:
```
benchmarking decode/en/aeson
time                 1.792 ms   (1.751 ms .. 1.849 ms)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 1.777 ms   (1.759 ms .. 1.799 ms)
std dev              65.93 μs   (54.43 μs .. 81.57 μs)
variance introduced by outliers: 23% (moderately inflated)

benchmarking decode/jp/aeson
time                 2.533 ms   (2.479 ms .. 2.603 ms)
                     0.995 R²   (0.993 R² .. 0.997 R²)
mean                 2.565 ms   (2.528 ms .. 2.613 ms)
std dev              133.9 μs   (106.6 μs .. 169.1 μs)
variance introduced by outliers: 34% (moderately inflated)

benchmarking encode/en/aeson-to-bytestring
time                 413.5 μs   (407.1 μs .. 418.3 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 412.5 μs   (406.9 μs .. 421.3 μs)
std dev              22.61 μs   (15.53 μs .. 33.70 μs)
variance introduced by outliers: 50% (moderately inflated)

benchmarking encode/en/aeson-via-text-to-bytestring
time                 1.115 ms   (1.092 ms .. 1.138 ms)
                     0.997 R²   (0.996 R² .. 0.998 R²)
mean                 1.098 ms   (1.085 ms .. 1.115 ms)
std dev              49.12 μs   (41.44 μs .. 57.85 μs)
variance introduced by outliers: 33% (moderately inflated)

benchmarking encode/en/aeson-to-text
time                 909.1 μs   (895.2 μs .. 928.1 μs)
                     0.996 R²   (0.994 R² .. 0.998 R²)
mean                 938.7 μs   (922.2 μs .. 956.4 μs)
std dev              59.11 μs   (50.91 μs .. 70.45 μs)
variance introduced by outliers: 51% (severely inflated)

benchmarking encode/jp/aeson-to-bytestring
time                 424.9 μs   (418.2 μs .. 432.7 μs)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 435.4 μs   (427.8 μs .. 448.3 μs)
std dev              31.43 μs   (20.61 μs .. 52.74 μs)
variance introduced by outliers: 63% (severely inflated)

benchmarking encode/jp/aeson-via-text-to-bytestring
time                 1.103 ms   (1.082 ms .. 1.126 ms)
                     0.997 R²   (0.996 R² .. 0.999 R²)
mean                 1.108 ms   (1.092 ms .. 1.130 ms)
std dev              62.74 μs   (48.19 μs .. 89.31 μs)
variance introduced by outliers: 45% (moderately inflated)

benchmarking encode/jp/aeson-to-text
time                 929.5 μs   (915.1 μs .. 943.1 μs)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 924.3 μs   (911.5 μs .. 944.5 μs)
std dev              52.88 μs   (34.71 μs .. 74.71 μs)
variance introduced by outliers: 47% (moderately inflated)
```