@maximkulkin 2015-05-26
I have addressed all issues except for:
* `JsonPath` and `JsonPathStep` were renamed to `JSONPath` and `JSONPathElement`. I feel like stripping JSON prefix would make JSONPath look like a regular file system path that would confuse users.
* `parseJSONIndexed` and `parseJSONElemAtIndex` I left unchanged. The reason for JSON infix there is to mimic regular `parseJSON` (which in fact they are with a little additions)

Also I have added "Since 0.9" documentation to `<?>` function. It needed such annotation but real version should be tweaked.

Benchmarks show a slight (IMO insignificant) slow down on benchmarks. Assuming the benefits this patch provides I think it is worth it.